+++ src/components Proflie/profile.js
@@ -2,6 +2,7 @@ import {useDispatch, useSelector, shallowEqual} from "react-redux";
 import {changeName, changeShowName} from "../store/profile/actions";
 import {Form} from "../components form/form";
 import {selectName, selectShowName} from "../store/profile/selectors";
+import {Presentation_profile} from "./Presentation_profile";
 
 
 export const Profile = () => {
@@ -20,17 +21,25 @@ export const Profile = () => {
 
     return (
         <>
-            <h1>Profile</h1>
             <Form onSubmit={handleChangeName} />
-            <div>
-                <input
-                    type="checkbox"
-                    checked={showName}
-                    value={showName}
-                    onChange={toggleShowName}
-                />Show Name
-                {showName && <p>{name}</p>}
-            </div>
+            <Presentation_profile
+                name={name}
+                showName={showName}
+                toggleShowName={toggleShowName}
+            />
         </>
+        // <>
+        //     <h1>Profile</h1>
+        //     <Form onSubmit={handleChangeName} />
+        //     <div>
+        //         <input
+        //             type="checkbox"
+        //             checked={showName}
+        //             value={showName}
+        //             onChange={toggleShowName}
+        //         />Show Name
+        //         {showName && <p>{name}</p>}
+        //     </div>
+        // </>
     );
 };
