+++ src/store/index.js
@@ -1,7 +1,12 @@
-import {createStore, combineReducers} from 'redux';
+import {createStore, combineReducers, compose, applyMiddleware} from 'redux';
 import {profileReducer} from "./profile/reducer";
 import {chatsReducer} from "./chats/reducer";
 import {messagesReducer} from "./messages/reducer";
+import thunk from "redux-thunk";
+import {persistReducer, persistStore} from "redux-persist";
+import storage from 'redux-persist/lib/storage';
+
+const composeEnhansers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
 
 const rootReducer = combineReducers({
     profile: profileReducer,
@@ -9,8 +14,16 @@ const rootReducer = combineReducers({
     messages: messagesReducer,
 });
 
+const persistConfig = {
+    key: 'reactMessenger',
+    storage,
+}
+
+const persistedReducer = persistReducer(persistConfig, rootReducer);
+
 export const store = createStore(
-    rootReducer,
-    window.__REDUX_DEVTOOLS_EXTENSION__ &&
-    window.__REDUX_DEVTOOLS_EXTENSION__(),
+    persistedReducer,
+    composeEnhansers(applyMiddleware(thunk)),
 );
+
+export const persistor = persistStore(store);
