+++ src/store/messages/actions.js
@@ -1,3 +1,5 @@
+import {AUTHORS} from "../../Utils/constants";
+
 export const ADD_MESSAGE = 'MESSAGES::ADD_MESSAGE';
 export const DELETE_MESSAGE = 'MESSAGES::DELETE_MESSAGE';
 export const INIT_MESSAGES_BY_CHAT_ID = 'MESSAGES::INIT_MESSAGES_BY_CHAT_ID';
@@ -23,4 +25,21 @@ export const addMessage = (id, text) => ({
         text,
     },
 });
+
+let timeout;
+
+export const addMessageWithThunk = (id, text) => (dispatch, getState) => {
+    dispatch(addMessage(id, text));
+
+    if (text.author !== AUTHORS.BOT) {
+        clearTimeout(timeout);
+        timeout = setTimeout(() => {
+            const newMsg = {
+                value: 'Hello! You entered message',
+                author: AUTHORS.BOT,
+            };
+            dispatch(addMessage(id, newMsg));
+        }, 3000);
+    };
+};
